Rename from bulk_validator to validator
This tool can also be used to validate input files for live loader.
