+++ dgraph/cmd/root.go
@@ -22,10 +22,10 @@ import (
 	"strings"
 
 	"github.com/dgraph-io/dgraph/dgraph/cmd/migrate"
+	"github.com/dgraph-io/dgraph/dgraph/cmd/validator"
 
 	"github.com/dgraph-io/dgraph/dgraph/cmd/alpha"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/bulk"
-	bulkvalidator "github.com/dgraph-io/dgraph/dgraph/cmd/bulk_validator"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/cert"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/conv"
 	"github.com/dgraph-io/dgraph/dgraph/cmd/counter"
@@ -74,7 +74,7 @@ var rootConf = viper.New()
 // subcommands initially contains all default sub-commands.
 var subcommands = []*x.SubCommand{
 	&bulk.Bulk, &cert.Cert, &conv.Conv, &live.Live, &alpha.Alpha, &zero.Zero, &version.Version,
-	&debug.Debug, &counter.Increment, &migrate.Migrate, &bulkvalidator.BulkValidator,
+	&debug.Debug, &counter.Increment, &migrate.Migrate, &validator.Validator,
 }
 
 func initCmds() {
