+++ dgraph/cmd/validator/loader.go
@@ -1,4 +1,4 @@
-package bulkvalidator
+package validator
 
 import (
 	"bytes"
